legacy/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
Pillow>=10.0.0
moviepy>=1.0.3
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
facebook-sdk>=3.1.0
python-telegram-bot>=20.0
colorama>=0.4.6
rich>=13.0.0
pydantic>=2.0.0
pyyaml>=6.0.0
pytz>=2023.3
mysql-connector-python>=8.0.0

# Voice changing dependencies (so-vits-svc)
torch>=2.0.0
torchaudio>=2.0.0
soundfile>=0.12.1
numpy>=1.24.0
librosa>=0.10.0
scipy>=1.10.0
faiss-cpu>=1.7.4
praat-parselmouth>=0.4.3
ffmpeg-python>=0.2.0
resampy>=0.4.2
pyworld>=0.3.2
torchcrepe>=0.0.20

# Voice conversion dependencies
onnxruntime>=1.15.0
huggingface-hub>=0.16.0
pydub>=0.25.1

# Silero TTS for natural Russian voice
omegaconf>=2.3.0

# Whisper for speech-to-text transcription
openai-whisper>=20231117
tiktoken>=0.5.0

# Russian accent/stress marking for proper pronunciation
russtress>=0.1.0

# Audio separation (voice/background separation)
audio-separator>=0.17.0